Updated module augmentation for select-all package.
