+++ packages/ckeditor5-select-all/src/selectallcommand.ts
@@ -72,9 +72,3 @@ export default class SelectAllCommand extends Command {
 function isSelectAllScope( schema: Schema, element: Element ): boolean {
 	return schema.isLimit( element ) && ( schema.checkChild( element, '$text' ) || schema.checkChild( element, 'paragraph' ) );
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface CommandsMap {
-		selectAll: SelectAllCommand;
-	}
-}
