+++ packages/ckeditor5-select-all/src/selectallui.ts
@@ -57,9 +57,3 @@ export default class SelectAllUI extends Plugin {
 		} );
 	}
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface PluginsMap {
-		[ SelectAllUI.pluginName ]: SelectAllUI;
-	}
-}
